# TaxiCompany/BlazorApp1/WebApi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TaxiCompany.WebApi",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Car/driver/{driverId}": {
      "get": {
        "tags": [
          "Car"
        ],
        "summary": "Получает сведения о конкретном водителе и его автомобиле.",
        "parameters": [
          {
            "name": "driverId",
            "in": "path",
            "description": "Идентификатор водителя.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DriverCarInfoDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverCarInfoDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverCarInfoDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Driver": {
      "get": {
        "tags": [
          "Driver"
        ],
        "summary": "Получает список всех водителей.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Driver"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Driver"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Driver"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Driver"
        ],
        "summary": "Добавляет нового водителя.",
        "requestBody": {
          "description": "Объект водителя, который нужно добавить.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DriverDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Driver/{id}": {
      "get": {
        "tags": [
          "Driver"
        ],
        "summary": "Получает водителя по идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор водителя.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Driver"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Driver"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Driver"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Driver"
        ],
        "summary": "Обновляет данные водителя по идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор водителя, данные которого нужно обновить.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Объект водителя с новыми данными.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DriverDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Driver"
        ],
        "summary": "Удаляет водителя по идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор водителя, которого нужно удалить.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CarDto": {
        "required": [
          "assignedDriverId",
          "colour",
          "model",
          "realeseYear",
          "serialNumber"
        ],
        "type": "object",
        "properties": {
          "colour": {
            "type": "string",
            "description": "Цвет",
            "nullable": true
          },
          "model": {
            "type": "string",
            "description": "Модель",
            "nullable": true
          },
          "serialNumber": {
            "type": "string",
            "description": "Серийный номер",
            "nullable": true
          },
          "realeseYear": {
            "type": "string",
            "description": "Год выпуска",
            "format": "date-time"
          },
          "assignedDriverId": {
            "type": "integer",
            "description": "Идентификатор водителя",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "Driver": {
        "required": [
          "address",
          "assignedCarId",
          "fullName",
          "id",
          "passport",
          "phoneNumber"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "fullName": {
            "maxLength": 50,
            "minLength": 1,
            "type": "string"
          },
          "phoneNumber": {
            "maxLength": 11,
            "minLength": 1,
            "type": "string"
          },
          "passport": {
            "maxLength": 10,
            "minLength": 1,
            "type": "string"
          },
          "address": {
            "maxLength": 80,
            "minLength": 1,
            "type": "string"
          },
          "assignedCarId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "DriverCarInfoDto": {
        "type": "object",
        "properties": {
          "driver": {
            "$ref": "#/components/schemas/DriverDto"
          },
          "car": {
            "$ref": "#/components/schemas/CarDto"
          }
        },
        "additionalProperties": false
      },
      "DriverDto": {
        "required": [
          "address",
          "fullName",
          "passport",
          "phoneNumber"
        ],
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "description": "Имя и фамилия",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "description": "Номер телефона",
            "nullable": true
          },
          "passport": {
            "type": "string",
            "description": "Паспортные данные",
            "nullable": true
          },
          "address": {
            "type": "string",
            "description": "Адрес",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}